Removed video dependicies + migrated code to fabric_canvas
